+++ controller/src/main/java/com/linbit/linstor/api/rest/v1/Actions.java
@@ -76,7 +76,8 @@ public class Actions
             List<SnapReq> createSnapReqs = new ArrayList<>();
             for (Snapshot snapReq : req.snapshots)
             {
-                createSnapReqs.add(new SnapReq(snapReq.nodes, snapReq.resource_name, snapReq.name));
+                createSnapReqs.add(new SnapReq(
+                    snapReq.nodes, snapReq.resource_name, snapReq.name, snapReq.snapshot_definition_props));
             }
             Flux<ApiCallRc> flux = snapCrtHandler.createSnapshot(new CreateMultiSnapRequest(createSnapReqs));