+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow setting properties on snapshot definitions
+
 ### Changed
 
 - LVM-Thin snapshots are set to read-only now